requirements.txt
# Core dependencies for Wisteria CTR Studio
openai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# Production server
gunicorn>=21.0.0

# Google Cloud Storage (optional)
google-cloud-storage>=2.10.0

# Google Gemini AI
google-generativeai>=0.3.0

# HTTP client dependencies
requests>=2.28.0
aiohttp>=3.9.0

# RunPod SDK for pod management
runpod>=1.6.0

# Optional dependencies (uncomment as needed)
# pytest>=7.0.0     # For testing
# pytest-asyncio   # For async testing